+++ cmd/pgp-tomb/pgp-tomb.go
@@ -3,6 +3,7 @@ package main
 import (
 	"fmt"
 	"os"
+	"os/exec"
 	"strings"
 	"syscall"
 
@@ -76,6 +77,10 @@ var (
 		BashCompletionFunction: bashCompletionFunction,
 		PersistentPreRun: func(cmd *cobra.Command, args []string) {
 			initConfig()
+			executeHook("pre", cmd.Name())
+		},
+		PersistentPostRun: func(cmd *cobra.Command, args []string) {
+			executeHook("post", cmd.Name())
 		},
 	}
 )
@@ -121,6 +126,31 @@ func initConfig() {
 
 	// Validate & initialize configuration.
 	config.Init()
+}
+
+func executeHook(alias string, command string) {
+	hooks := config.GetHooks()
+	if hook, found := hooks[alias]; found {
+		args := []string{
+			config.GetRoot(),
+			command,
+		}
+		cmd := exec.Command(hook.Path, args...)
+		cmd.Stdout = os.Stdout
+		cmd.Stderr = os.Stderr
+		if err := cmd.Run(); err != nil {
+			switch err := err.(type) {
+			case *exec.ExitError:
+				fmt.Fprintf(os.Stderr, "Aborted by '%s' hook!\n", alias)
+				os.Exit(1)
+			default:
+				logrus.WithFields(logrus.Fields{
+					"error": err,
+					"hook":  alias,
+				}).Fatal("Failed to execute hook!")
+			}
+		}
+	}
 }
 
 func parseTags(tags []string) []secret.Tag {
@@ -333,10 +363,13 @@ func main() {
 		Short: "Initialize a new tomb",
 		Args:  cobra.ExactArgs(1),
 		PersistentPreRun: func(cmd *cobra.Command, args []string) {
-			// Skip initConfig().
+			// Skip initialization of configuration & execution of hooks.
 		},
 		Run: func(cmd *cobra.Command, args []string) {
 			core.Initialize(args[0])
+		},
+		PersistentPostRun: func(cmd *cobra.Command, args []string) {
+			// Skip execution of hooks.
 		},
 	}
 
@@ -346,10 +379,13 @@ func main() {
 		Short: "Generate Bash completion script",
 		Args:  cobra.NoArgs,
 		PersistentPreRun: func(cmd *cobra.Command, args []string) {
-			// Skip initConfig().
+			// Skip initialization of configuration & execution of hooks.
 		},
 		Run: func(cmd *cobra.Command, args []string) {
 			rootCmd.GenBashCompletion(os.Stdout)
+		},
+		PersistentPostRun: func(cmd *cobra.Command, args []string) {
+			// Skip execution of hooks.
 		},
 	}
 
@@ -359,13 +395,16 @@ func main() {
 		Short: "Generate Zsh completion script",
 		Args:  cobra.NoArgs,
 		PersistentPreRun: func(cmd *cobra.Command, args []string) {
-			// Skip initConfig().
+			// Skip initialization of configuration & execution of hooks.
 		},
 		Run: func(cmd *cobra.Command, args []string) {
 			// XXX: custom completion function for secret URIs not yet available
 			// for Zsh. See:
 			//   - https://github.com/spf13/cobra/pull/884
 			rootCmd.GenZshCompletion(os.Stdout)
+		},
+		PersistentPostRun: func(cmd *cobra.Command, args []string) {
+			// Skip execution of hooks.
 		},
 	}